requirements.txt
aniso8601==9.0.1
click==8.1.3
Flask==2.1.2
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
gevent==21.12.0
greenlet==1.1.2
importlib-metadata==4.11.3
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==1.4.2
PyMySQL==1.0.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
